+++ Package.swift
@@ -13,7 +13,7 @@
 //  `swift test`
 //
 //  Created by Matthias Zenger on 09/08/2019.
-//  Copyright © 2019-2021 Google LLC.
+//  Copyright © 2019-2022 Google LLC.
 //
 //  Licensed under the Apache License, Version 2.0 (the "License");
 //  you may not use this file except in compliance with the License.